Good reorganisation of lisa.utils, comments.
