+++ lisa-theories/src/main/scala/lisa/settheory/SetTheoryTGAxioms.scala
@@ -1,7 +1,7 @@
 package lisa.settheory
 
 import lisa.kernel.fol.FOL.*
-import lisa.utils.Helpers.{_, given}
+import lisa.utils.KernelHelpers.{_, given}
 
 /**
  * Axioms for the Tarski-Grothendieck theory (TG)
