--- lisa-utils/src/main/scala/lisa/utils/Helpers.scala
@@ -1,10 +1,0 @@
-package lisa.utils
-
-/**
- * A helper file that provides various syntactic sugars for LISA's FOL and proofs.
- * Usage:
- * <pre>
- * import utilities.Helpers.*
- * </pre>
- */
-object Helpers extends TheoriesHelpers
