+++ lisa-utils/src/test/scala/lisa/utils/ParserTest.scala
@@ -3,10 +3,10 @@ package lisa.utils
 import lisa.kernel.fol.FOL._
 import lisa.kernel.proof.SequentCalculus.Sequent
 import lisa.utils.FOLParser
-import lisa.utils.Helpers.*
-import lisa.utils.Helpers._
-import lisa.utils.Helpers.given_Conversion_Identifier_String
-import lisa.utils.Helpers.given_Conversion_String_Identifier
+import lisa.utils.KernelHelpers.*
+import lisa.utils.KernelHelpers._
+import lisa.utils.KernelHelpers.given_Conversion_Identifier_String
+import lisa.utils.KernelHelpers.given_Conversion_String_Identifier
 import org.scalatest.funsuite.AnyFunSuite
 
 class ParserTest extends AnyFunSuite with TestUtils {
