+++ src/main/scala/lisa/automation/kernel/OLPropositionalSolver.scala
@@ -3,13 +3,9 @@ package lisa.automation.kernel
 import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.*
-import lisa.kernel.proof.SequentCalculus as LK
-import lisa.utils.Helpers.{_, given}
-import lisa.utils.Helpers.{_, given}
-import lisa.utils.Library
-import lisa.utils.tactics.ProofTacticLib.ParameterlessAndThen
-import lisa.utils.tactics.ProofTacticLib.ParameterlessHave
-import lisa.utils.tactics.ProofTacticLib.ProofTactic
+import lisa.utils.KernelHelpers.{_, given}
+import lisa.prooflib.Library
+import lisa.prooflib.ProofTacticLib.*
 
 object OLPropositionalSolver {
