+++ src/main/scala/lisa/automation/kernel/SimplePropositionalSolver.scala
@@ -4,11 +4,9 @@ import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.*
 import lisa.utils.FOLPrinter
-import lisa.utils.Helpers.{_, given}
-import lisa.utils.Library
-import lisa.utils.tactics.ProofTacticLib.ParameterlessAndThen
-import lisa.utils.tactics.ProofTacticLib.ParameterlessHave
-import lisa.utils.tactics.ProofTacticLib.ProofTactic
+import lisa.utils.KernelHelpers.{_, given}
+import lisa.prooflib.Library
+import lisa.prooflib.ProofTacticLib.*
 
 import scala.collection.mutable.Set as mSet
