+++ src/main/scala/lisa/proven/peano_example/PeanoArithmeticsLibrary.scala
@@ -1,5 +1,5 @@
 package lisa.proven.peano_example
 
-object PeanoArithmeticsLibrary extends lisa.utils.Library(lisa.proven.peano_example.PeanoArithmetics.runningPeanoTheory) {
+object PeanoArithmeticsLibrary extends lisa.prooflib.Library(lisa.proven.peano_example.PeanoArithmetics.runningPeanoTheory) {
   export lisa.proven.peano_example.PeanoArithmetics.*
 }
