+++ src/test/scala/lisa/utilities/Transformations.scala
@@ -6,7 +6,7 @@ import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.*
 import lisa.test.ProofCheckerSuite
-import lisa.utils.Helpers.{_, given}
+import lisa.utils.KernelHelpers.{_, given}
 import lisa.utils.Printer
 
 class Transformations extends ProofCheckerSuite {
